+++ featuretools/tests/plugin_tests/featuretools_plugin/setup.py
