+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 description-file = README.md
 [tool:pytest]
-addopts = --doctest-modules
+addopts = --doctest-modules --ignore=featuretools/tests/plugin_tests/featuretools_plugin -k test_plugin
 python_files = featuretools/tests/*
 filterwarnings =
     ignore::DeprecationWarning
